Import a new version of XROOTD (20090821-1000):
    - Fixes in bulk prepare and sync readv operations
    - Add support for 'make install' / 'make uninstall' and other improvements in configure.classic
    - Several improvements / fixes in xrdcp:
      - reduced memory and CPU consumption;
      - extreme cp optimizations;
      - windows porting
    - Support for if/else if/else/fi constructs
    - Several portability fixes


git-svn-id: http://root.cern.ch/svn/root/trunk@29874 27541ba8-7e3a-0410-8455-c3a389f83636

+++ net/xrootd/src/xrootd/src/XrdCms/XrdCmsConfig.cc
@@ -906,7 +906,8 @@ int XrdCmsConfig::PidFile()
     else    snprintf(pidFN, sizeof(pidFN), "%s/cmsd.mangr.pid", ppath);
 
     if ((xfd = open(pidFN, O_WRONLY|O_CREAT|O_TRUNC,0644)) < 0) xop = "open";
-       else {if ((write(xfd,buff,snprintf(buff,sizeof(buff),"%d",getpid()))<0)
+       else {if ((write(xfd,buff,snprintf(buff,sizeof(buff),"%d",
+                            static_cast<int>(getpid()))) < 0)
              || (LocalRoot && 
                            (write(xfd,(void *)"\n&pfx=",6)  < 0 ||
                             write(xfd,(void *)LocalRoot,strlen(LocalRoot)) < 0
